+++ CHANGES.rst
@@ -8,6 +8,7 @@
 * New UserSelectField and UserSelectMultiField for Lastuser autocomplete
 * Revamped forms.html and autoform generator
 * ValidEmailDomain validator is now ValidEmail and does syntax checks
+* New NoObfuscatedEmail validator looks for obfuscated emails
 
 
 0.3.0